requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
python-dotenv>=1.0.0
validators>=0.22.0
lxml>=5.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
